Use Int64 for password_resets user_id

+++ src/browser_authentication_app_skeleton/src/pages/password_resets/new_page.cr
@@ -1,6 +1,6 @@
 class PasswordResets::NewPage < AuthLayout
   needs operation : ResetPassword
-  needs user_id : Int32
+  needs user_id : Int64
 
   def content
     h1 "Reset your password"
